Python/requirements.txt
fastapi
uvicorn
psycopg2-binary
pgvector
python-dotenv
langchain
langchain-google-genai
langchain-postgres
google-generativeai
langchain-chroma
chromadb
